licenses and abstractions
