+++ Utilities.Rest/Packages/com.utilities.rest/Runtime/AbstractAuthentication.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9b50a78f4e0220045b408adfa46d340d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
