+++ Utilities.Rest/Packages/com.utilities.rest/Runtime/BaseEndPoint.cs
@@ -1,3 +1,5 @@
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
 using Utilities.WebRequestRest.Interfaces;
 
 namespace Utilities.WebRequestRest
