+++ Utilities.Rest/Packages/com.utilities.rest/Runtime/Extensions/StringExtensions.cs
@@ -1,3 +1,5 @@
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
 using System.Net.Http;
 using System.Text;
