+++ Utilities.Rest/Packages/com.utilities.rest/Runtime/Interfaces/IAuthInfo.cs
@@ -1,3 +1,5 @@
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
 namespace Utilities.WebRequestRest.Interfaces
 {
     public interface IAuthInfo
